テストのための補助関数をすべてutils_test.goに移動させた 機能の重複していたgenerateMassageをmakeMessageに統一 generate->makeに変更
